requirements.txt
fastapi
uvicorn
faster-whisper
torch
torchaudio
webrtcvad
transformers
spacy
language-tool-python
numpy
scikit-learn
xgboost
request